voice-ledger/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.27
pydantic==2.6.3
python-dotenv==1.0.1

# ── AI — Google Gemini only (FREE, just needs Gmail) ───────────────────────────
google-generativeai==0.5.0
langchain-google-genai==1.0.1
langchain-core==0.1.41

# ── Observability (skinny version, ~10x smaller than full mlflow) ──────────────
mlflow-skinny==2.11.1
